Move assertHtml() tests into a separate test case.
Don't indirectly tests via a 'fixture' that is also a test case. This
pattern is kind of confusing and having assertHtml() be its own test
case feels simpler.
