+++ tests/TestCase/TestSuite/TestCaseTest.php
@@ -24,7 +24,6 @@ use Cake\ORM\Entity;
 use Cake\ORM\Table;
 use Cake\ORM\TableRegistry;
 use Cake\TestSuite\TestCase;
-use Cake\Test\Fixture\AssertHtmlTestCase;
 use Cake\Test\Fixture\FixturizedTestCase;
 
 /**
@@ -120,135 +119,6 @@ class TestCaseTest extends TestCase
         $event = new Event('my.event');
         $manager->dispatch($event);
         $this->assertEventFired('my.event', $manager);
-    }
-
-    /**
-     * testAssertHtml
-     *
-     * @return void
-     */
-    public function testAssertHtmlBasic()
-    {
-        $test = new AssertHtmlTestCase('testAssertHtmlQuotes');
-        $result = $test->run();
-
-        $this->assertEquals(0, $result->errorCount());
-        $this->assertTrue($result->wasSuccessful());
-        $this->assertEquals(0, $result->failureCount());
-    }
-
-    /**
-     * test assertHtml works with single and double quotes
-     *
-     * @return void
-     */
-    public function testAssertHtmlQuoting()
-    {
-        $input = '<a href="/test.html" class="active">My link</a>';
-        $pattern = [
-            'a' => ['href' => '/test.html', 'class' => 'active'],
-            'My link',
-            '/a'
-        ];
-        $this->assertHtml($pattern, $input);
-
-        $input = "<a href='/test.html' class='active'>My link</a>";
-        $pattern = [
-            'a' => ['href' => '/test.html', 'class' => 'active'],
-            'My link',
-            '/a'
-        ];
-        $this->assertHtml($pattern, $input);
-
-        $input = "<a href='/test.html' class='active'>My link</a>";
-        $pattern = [
-            'a' => ['href' => 'preg:/.*\.html/', 'class' => 'active'],
-            'My link',
-            '/a'
-        ];
-        $this->assertHtml($pattern, $input);
-
-        $input = "<span><strong>Text</strong></span>";
-        $pattern = [
-            '<span',
-            '<strong',
-            'Text',
-            '/strong',
-            '/span'
-        ];
-        $this->assertHtml($pattern, $input);
-
-        $input = "<span class='active'><strong>Text</strong></span>";
-        $pattern = [
-            'span' => ['class'],
-            '<strong',
-            'Text',
-            '/strong',
-            '/span'
-        ];
-        $this->assertHtml($pattern, $input);
-    }
-
-    /**
-     * Test that assertHtml runs quickly.
-     *
-     * @return void
-     */
-    public function testAssertHtmlRuntimeComplexity()
-    {
-        $pattern = [
-            'div' => [
-                'attr1' => 'val1',
-                'attr2' => 'val2',
-                'attr3' => 'val3',
-                'attr4' => 'val4',
-                'attr5' => 'val5',
-                'attr6' => 'val6',
-                'attr7' => 'val7',
-                'attr8' => 'val8',
-            ],
-            'My div',
-            '/div'
-        ];
-        $input = '<div attr8="val8" attr6="val6" attr4="val4" attr2="val2"' .
-            ' attr1="val1" attr3="val3" attr5="val5" attr7="val7" />' .
-            'My div' .
-            '</div>';
-        $this->assertHtml($pattern, $input);
-    }
-
-    /**
-     * testNumericValuesInExpectationForAssertHtml
-     *
-     * @return void
-     */
-    public function testNumericValuesInExpectationForAssertHtml()
-    {
-        $test = new AssertHtmlTestCase('testNumericValuesInExpectationForAssertHtml');
-        $result = $test->run();
-        $this->assertEquals(0, $result->errorCount());
-        $this->assertTrue($result->wasSuccessful());
-        $this->assertEquals(0, $result->failureCount());
-    }
-
-    /**
-     * testBadAssertHtml
-     *
-     * @return void
-     */
-    public function testBadAssertHtml()
-    {
-        $test = new AssertHtmlTestCase('testBadAssertHtml');
-        $result = $test->run();
-        $this->assertEquals(0, $result->errorCount());
-        $this->assertFalse($result->wasSuccessful());
-        $this->assertEquals(1, $result->failureCount());
-
-        $test = new AssertHtmlTestCase('testBadAssertHtml2');
-        $result = $test->run();
-        $this->assertEquals(0, $result->errorCount());
-        $this->assertFalse($result->wasSuccessful());
-        $this->assertEquals(1, $result->failureCount());
     }
 
     /**
